refactor(server): allow injecting logic for building package artifacts
To be able to mock it in tests.
