+++ jit-binding-server/src/main/kotlin/io/github/typesafegithub/workflows/jitbindingserver/Main.kt
@@ -7,7 +7,9 @@ import com.sksamuel.aedile.core.refreshAfterWrite
 import io.github.oshai.kotlinlogging.KotlinLogging.logger
 import io.github.typesafegithub.workflows.actionbindinggenerator.domain.ActionCoords
 import io.github.typesafegithub.workflows.mavenbinding.Artifact
+import io.github.typesafegithub.workflows.mavenbinding.buildPackageArtifacts
 import io.github.typesafegithub.workflows.mavenbinding.buildVersionArtifacts
+import io.github.typesafegithub.workflows.shared.internal.getGithubAuthToken
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.Application
 import io.ktor.server.application.ApplicationCall
@@ -48,19 +50,26 @@ fun main() {
         logger.error(throwable) { "Uncaught exception in thread $thread" }
     }
     embeddedServer(Netty, port = 8080) {
-        appModule(buildVersionArtifacts = ::buildVersionArtifacts)
+        appModule(
+            buildVersionArtifacts = ::buildVersionArtifacts,
+            buildPackageArtifacts = ::buildPackageArtifacts,
+        )
     }.start(wait = true)
 }
 
-fun Application.appModule(buildVersionArtifacts: (ActionCoords) -> Map<String, Artifact>?) {
+fun Application.appModule(
+    buildVersionArtifacts: (ActionCoords) -> Map<String, Artifact>?,
+    buildPackageArtifacts: suspend (ActionCoords, String, (Collection<ActionCoords>) -> Unit) -> Map<String, String>,
+) {
     val bindingsCache = buildBindingsCache(buildVersionArtifacts)
+    val metadataCache = buildMetadataCache(bindingsCache, buildPackageArtifacts)
     installPlugins(prometheusRegistry)
 
     routing {
         internalRoutes(prometheusRegistry)
 
         artifactRoutes(bindingsCache, prometheusRegistry)
-        metadataRoutes(bindingsCache, prometheusRegistry)
+        metadataRoutes(metadataCache, prometheusRegistry)
     }
 }
 
@@ -72,6 +81,25 @@ private fun buildBindingsCache(
         .refreshAfterWrite(1.hours)
         .recordStats()
         .asLoadingCache<ActionCoords, ArtifactResult> { runCatching { buildVersionArtifacts(it) } }
+
+@Suppress("ktlint:standard:function-signature") // Conflict with detekt.
+private fun buildMetadataCache(
+    bindingsCache: LoadingCache<ActionCoords, ArtifactResult>,
+    buildPackageArtifacts: suspend (ActionCoords, String, (Collection<ActionCoords>) -> Unit) -> Map<String, String>,
+): LoadingCache<ActionCoords, MetadataResult> =
+    Caffeine
+        .newBuilder()
+        .refreshAfterWrite(1.hours)
+        .recordStats()
+        .asLoadingCache<ActionCoords, MetadataResult> {
+            runCatching {
+                buildPackageArtifacts(
+                    it,
+                    getGithubAuthToken(),
+                    { coords -> prefetchBindingArtifacts(coords, bindingsCache) },
+                )
+            }
+        }
 
 val deliverOnRefreshRoute = System.getenv("GWKT_DELIVER_ON_REFRESH").toBoolean()
