+++ jit-binding-server/src/test/kotlin/io/github/typesafegithub/workflows/jitbindingserver/ArtifactRoutesTest.kt
@@ -23,6 +23,8 @@ class ArtifactRoutesTest :
                             buildVersionArtifacts = {
                                 mapOf("some-action-v4.pom" to TextArtifact { "Some POM contents" })
                             },
+                            // Irrelevant for these tests.
+                            buildPackageArtifacts = { _, _, _ -> emptyMap() },
                         )
                     }
 
@@ -41,6 +43,8 @@ class ArtifactRoutesTest :
                     application {
                         appModule(
                             buildVersionArtifacts = { null },
+                            // Irrelevant for these tests.
+                            buildPackageArtifacts = { _, _, _ -> emptyMap() },
                         )
                     }
 
@@ -58,6 +62,8 @@ class ArtifactRoutesTest :
                     application {
                         appModule(
                             buildVersionArtifacts = { error("An internal error occurred!") },
+                            // Irrelevant for these tests.
+                            buildPackageArtifacts = { _, _, _ -> emptyMap() },
                         )
                     }
 
@@ -79,6 +85,8 @@ class ArtifactRoutesTest :
                     application {
                         appModule(
                             buildVersionArtifacts = mockBuildVersionArtifacts,
+                            // Irrelevant for these tests.
+                            buildPackageArtifacts = { _, _, _ -> emptyMap() },
                         )
                     }
